fix::Refactor getTablesDefinition method (Issue #23) 101a
